folder restructure + added blinky demo
