+++ main/component.mk
@@ -2,5 +2,5 @@
 # "main" pseudo-component makefile.
 #
 # (Uses default behaviour of compiling all source files in directory, adding 'include' to include path.)
-COMPONENT_SRCDIRS = . drivers drivers/com
-COMPONENT_ADD_INCLUDEDIRS = drivers drivers/com
+COMPONENT_SRCDIRS = . drivers drivers/com app
+COMPONENT_ADD_INCLUDEDIRS = include include/app include/drivers include/drivers/com
